# Avoid `ByteString` allocation during zstd decompression 