Also adds a benchmark for zstd compression/decompression performance that indicates a 1.3x increase in throughput.

Before:
```
Benchmark                  (size)   Mode  Cnt       Score      Error  Units
ZstdBenchmark.compress       4096  thrpt   25   25099.133 ±  190.562  ops/s
ZstdBenchmark.compress    4194304  thrpt   25     125.987 ±    0.159  ops/s
ZstdBenchmark.decompress     4096  thrpt   25  116896.309 ± 3719.310  ops/s
ZstdBenchmark.decompress  4194304  thrpt   25     919.009 ±    2.409  ops/s
```

After:
```
Benchmark                  (size)   Mode  Cnt       Score      Error  Units
ZstdBenchmark.compress       4096  thrpt   25   25083.031 ±  270.572  ops/s
ZstdBenchmark.compress    4194304  thrpt   25     125.919 ±    0.235  ops/s
ZstdBenchmark.decompress     4096  thrpt   25  147747.084 ± 4691.230  ops/s
ZstdBenchmark.decompress  4194304  thrpt   25    1325.584 ±    1.861  ops/s
```